selectActiveLabel, and removing generator code (#5374)
