+++ scripts/snapshots/es6Files.txt
@@ -114,6 +114,7 @@
   "es6/state/selectors/barSelectors.js",
   "es6/state/selectors/brushSelectors.js",
   "es6/state/selectors/combiners",
+  "es6/state/selectors/combiners/combineActiveLabel.js",
   "es6/state/selectors/combiners/combineAxisRangeWithReverse.js",
   "es6/state/selectors/containerSelectors.js",
   "es6/state/selectors/dataSelectors.js",
