+++ src/context/tooltipContext.tsx
@@ -1,46 +1,7 @@
 import React, { createContext, useContext } from 'react';
-import { ChartCoordinate, Coordinate, DataKey } from '../util/types';
+import { Coordinate, DataKey } from '../util/types';
 import { useAppDispatch } from '../state/hooks';
 import { mouseLeaveItem, setActiveClickItemIndex, setActiveMouseOverItemIndex } from '../state/tooltipSlice';
-
-export type TooltipContextValue = {
-  label: string;
-  payload: any[];
-  coordinate: ChartCoordinate;
-  active: boolean;
-  index: number;
-};
-
-export const doNotDisplayTooltip: TooltipContextValue = {
-  label: '',
-  payload: [],
-  coordinate: { x: 0, y: 0 },
-  active: false,
-  index: -1,
-};
-
-const TooltipContext = createContext<TooltipContextValue>(doNotDisplayTooltip);
-
-/**
- * @deprecated do not use; instead prefer actions from tooltipSlice
- *
- * This depends on state set from generateCategoricalChart.
- */
-export const TooltipContextProvider = TooltipContext.Provider;
-
-/**
- * @deprecated this is depending on generateCategoricalChart state, do not use.
- *
- * Instead, use:
- * - selectActiveIndex for index
- * - selectActiveCoordinate for coordinate
- * - selectActiveLabel for label
- * - selectTooltipPayload for payload
- * - selectIsTooltipActive for active
- *
- * @returns deprecated, do not use
- */
-export const useTooltipContext = () => useContext(TooltipContext);
 
 export type TooltipPayloadType = any[];
